+++ app/src/main/java/com/guzzardi/dictionaryapp/data/datasources/DataSource.kt
@@ -1,6 +1,6 @@
 package com.guzzardi.dictionaryapp.data.datasources
 
-import com.guzzardi.dictionaryapp.data.DefinitionItemDto
+import com.guzzardi.dictionaryapp.data.model.DefinitionItemDto
 import kotlinx.coroutines.flow.Flow
 
 interface DataSource {
