+++ app/src/main/java/com/guzzardi/dictionaryapp/presentation/views/definitions/DefinitionsAdapter.kt
@@ -2,7 +2,7 @@ package com.guzzardi.dictionaryapp.presentation.views.definitions
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.guzzardi.dictionaryapp.data.DefinitionItemDto
+import com.guzzardi.dictionaryapp.data.model.DefinitionItemDto
 
 class DefinitionsAdapter(private val definitions: List<DefinitionItemDto>) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
